[SKIN] [DIAG]#12 Use explicit cast from double to int when retrieving skin error codes from a yarp bottle. Fix C4244 on Windows VS compilation as raised by @pattacini.

+++ src/modules/skinManager/src/SkinDiagnosticsReadThread.cpp
@@ -107,7 +107,7 @@ void SkinDiagnosticsReadThread::run(void) {
 
         // Extract skin errors
         deque<bool> errorTaxels;   // FG: Set to true if the taxel returned an error
-        int errorCode = (*data)[3];
+        int errorCode = (int) (*data)[3];
         if (!(errorCode & SkinErrorCode::StatusOK)) {
             // An error occurred
             // Handle stuck taxel data
